Created and implemented the Boot class
Created and implemented the Boot class which contains the Boot runner method which runs all of the methods to be executed at boost.
